requirements.txt
matplotlib>=1.5.1
numpy>=1.14.0
scipy>=0.19.1
tensorflow-gpu>=1.4.1
wget>=3.2
theano
keras
intervaltree
kerosene
mimir
resampy
sklearn